requirements.txt
asn1crypto==0.24.0
astroid==2.0.4
autopep8==1.4.3
cachetools==3.0.0
certifi==2018.10.15
cffi==1.11.5
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
cryptography==2.4.1
Django==2.1.3
django-cors-headers==2.4.0
django-csp==3.4
django-storages==1.7.1
djangorestframework==3.9.0
djangorestframework-simplejwt==3.2.3
google-api-core==1.5.2
google-auth==1.6.1
google-cloud-core==0.28.1
google-cloud-storage==1.13.0
google-resumable-media==0.3.1
googleapis-common-protos==1.5.5
gunicorn==19.9.0
idna==2.7
isort==4.3.4
itypes==1.1.0
Jinja2==2.10
lazy-object-proxy==1.3.1
MarkupSafe==1.0
mccabe==0.6.1
prometheus-client==0.4.2
protobuf==3.6.1
pyasn1==0.4.4
pyasn1-modules==0.2.2
pycodestyle==2.4.0
pycparser==2.19
PyJWT==1.6.4
pylint==2.1.1
PyMySQL==0.9.2
pytz==2018.7
requests==2.20.1
rsa==4.0
six==1.11.0
uritemplate==3.0.0
urllib3==1.24.1
wrapt==1.10.11
